requirements.txt
Package         Version
--------------- -------
blinker         1.6.2
click           8.1.3
et-xmlfile      1.1.0
Flask           2.3.2
itsdangerous    2.1.2
Jinja2          3.1.2
MarkupSafe      2.1.3
numpy           1.25.0
pandas          2.0.2
pip             23.0.1
psycopg2-binary 2.9.6
python-dateutil 2.8.2
python-dotenv   1.0.0
pytz            2023.3
setuptools      66.1.1
six             1.16.0
tzdata          2023.3
Werkzeug        2.3.6
